Properly evict stale PipelineCache entries associated with deleted VezPipelines
Users wishing to reload piplines (e.g. shader reload) will immediately encounter a bug
wherein a compiled pipeline associated with a destroyed VezPipeline will (resurrect)
due to the hash not taking into account shader modules.

Instead of augmenting the hash further, this commit implements a more complete fix,
to wit hashes are stored on the vez::Pipeline so that on destruction, all pipelines
in the PipelineCache associated with the vez::Pipeline are destroyed as well.